[CORE-4969] Updated logic load new video from url
